refactor: restructure CharacterGraph - move types and hooks to proper dirs
