+++ src/hooks/useCharacterGraphResize.ts
