+++ src/types/index.ts
@@ -7,3 +7,4 @@ export * from "./character"; // includes Place, Item
 export * from "./foreshadowing";
 export * from "./api"; // API response types
 export * from "./document";
+export * from "./characterGraph"; // D3 그래프 타입
